fix(mariadb): migrate function to typescript
