+++ src/dialects/mariadb/query-generator-typescript.ts
@@ -1,0 +1,39 @@
+import { rejectInvalidOptions } from '../../utils/check';
+import { joinSQLFragments } from '../../utils/join-sql-fragments';
+import { generateIndexName } from '../../utils/string';
+import { REMOVE_INDEX_QUERY_SUPPORTABLE_OPTIONS } from '../abstract/query-generator-typescript';
+import type { RemoveIndexQueryOptions, TableNameOrModel } from '../abstract/query-generator-typescript';
+import { MySqlQueryGenerator } from '../mysql/query-generator.js';
+
+/**
+ * Temporary class to ease the TypeScript migration
+ */
+export class MariaDbQueryGeneratorTypeScript extends MySqlQueryGenerator {
+  removeIndexQuery(tableName: TableNameOrModel, indexNameOrAttributes: string | string[], options: RemoveIndexQueryOptions) {
+    if (options) {
+      rejectInvalidOptions(
+        'removeIndexQuery',
+        this.dialect.name,
+        REMOVE_INDEX_QUERY_SUPPORTABLE_OPTIONS,
+        new Set(['ifExists']),
+        options,
+      );
+    }
+
+    let indexName;
+    const table = this.extractTableDetails(tableName);
+    if (Array.isArray(indexNameOrAttributes)) {
+      indexName = generateIndexName(table, { fields: indexNameOrAttributes });
+    } else {
+      indexName = indexNameOrAttributes;
+    }
+
+    return joinSQLFragments([
+      'DROP INDEX',
+      options?.ifExists ? 'IF EXISTS' : '',
+      this.quoteIdentifier(indexName),
+      'ON',
+      this.quoteTable(tableName),
+    ]);
+  }
+}
