+++ src/dialects/mariadb/query-generator.js
@@ -2,14 +2,11 @@
 
 import { normalizeDataType } from '../abstract/data-types-utils';
 import { joinSQLFragments } from '../../utils/join-sql-fragments.js';
-import { rejectInvalidOptions } from '../../utils/check';
-import { generateIndexName } from '../../utils/string';
-import { REMOVE_INDEX_QUERY_SUPPORTABLE_OPTIONS } from '../abstract/query-generator-typescript';
+import { MariaDbQueryGeneratorTypeScript } from './query-generator-typescript';
 
-const { MySqlQueryGenerator } = require('../mysql/query-generator');
 const _ = require('lodash');
 
-export class MariaDbQueryGenerator extends MySqlQueryGenerator {
+export class MariaDbQueryGenerator extends MariaDbQueryGeneratorTypeScript {
 
   _getTechnicalSchemaNames() {
     return ['MYSQL', 'INFORMATION_SCHEMA', 'PERFORMANCE_SCHEMA', 'mysql', 'information_schema', 'performance_schema'];
@@ -48,34 +45,6 @@ export class MariaDbQueryGenerator extends MySqlQueryGenerator {
       ifExists,
       this.quoteIdentifier(attributeName),
       ';',
-    ]);
-  }
-
-  removeIndexQuery(tableName, indexNameOrAttributes, options) {
-    if (options) {
-      rejectInvalidOptions(
-        'removeIndexQuery',
-        this.dialect.name,
-        REMOVE_INDEX_QUERY_SUPPORTABLE_OPTIONS,
-        new Set(['ifExists']),
-        options,
-      );
-    }
-
-    let indexName;
-    const table = this.extractTableDetails(tableName);
-    if (Array.isArray(indexNameOrAttributes)) {
-      indexName = generateIndexName(table, { fields: indexNameOrAttributes });
-    } else {
-      indexName = indexNameOrAttributes;
-    }
-
-    return joinSQLFragments([
-      'DROP INDEX',
-      options?.ifExists ? 'IF EXISTS' : '',
-      this.quoteIdentifier(indexName),
-      'ON',
-      this.quoteTable(tableName),
     ]);
   }
